backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
python-dotenv==1.1.0
requests==2.32.4
python-multipart==0.0.20
sqlalchemy==1.4.53
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
